@@ -2,20 +2,21 @@ name: tentura
 description: 'Social network for communities'
 publish_to: 'none'
 
-version: 0.7.6
+version: 0.8.5
 
 environment:
   sdk: '>=3.5.3 <4.0.0'
-  flutter: ^3.24.3
+  flutter: ^3.24.4
 
 dependencies:
   flutter:
     sdk: flutter
   nil: ^1.1.1
+  uuid: ^4.4.2
   intl: ^0.19.0
-  get_it: ^7.7.0
+  get_it: ^8.0.2
   latlong2: ^0.9.1
-  injectable: ^2.4.4
+  injectable: ^2.5.0
   auto_route: ^9.2.2
   collection: ^1.18.0
   built_value: ^8.9.2
@@ -27,18 +28,22 @@ dependencies:
 
   # Platform related
   geocoding: ^3.0.0
-  share_plus: ^10.0.2
   geolocator: ^13.0.1
+  share_plus: ^10.1.1
   image_picker: ^1.1.2
-  mobile_scanner: ^5.2.3
+  share_handler: ^0.0.22
+  mobile_scanner: ^6.0.2
   flutter_secure_storage: ^9.2.2
 
   # UI related
   readmore: ^3.0.0
+  qr_flutter: ^4.1.0
   flutter_map: ^7.0.2
-  flutter_svg: ^2.0.10
-  pretty_qr_code: ^3.3.0
-  vector_graphics: ^1.1.11
+  flutter_svg: ^2.0.13
+  google_fonts: ^6.2.1
+  vector_graphics: ^1.1.14
+  flutter_chat_ui: ^1.6.15
+  flutter_chat_types: ^3.6.2
   cached_network_image: ^3.4.1
   flutter_native_splash: ^2.4.0
   # smooth_page_indicator: ^1.2.0
@@ -48,8 +53,6 @@ dependencies:
   # Local
   tentura_sdk:
     path: packages/tentura_sdk
-  share_handler_multi:
-    path: packages/share_handler_multi
 
 dependency_overrides:
 
@@ -61,6 +64,7 @@ dev_dependencies:
   ferry_generator: ^0.12.0
   json_serializable: ^6.8.0
   very_good_analysis: ^6.0.0
+  build_web_compilers: ^4.0.11
   auto_route_generator: ^9.0.0
   injectable_generator: ^2.6.2
   flutter_launcher_icons: ^0.14.0
