@@ -1,9 +1,10 @@
+import 'package:flutter/foundation.dart';
 import 'package:injectable/injectable.dart';
 
 import 'package:tentura/consts.dart';
 import 'package:tentura_sdk/tentura_sdk.dart';
 
-export 'package:tentura_sdk/tentura_sdk.dart' show DataSource;
+export 'package:tentura_sdk/tentura_sdk.dart';
 
 @singleton
 class RemoteApiService extends TenturaApi {
@@ -15,9 +16,10 @@ class RemoteApiService extends TenturaApi {
   }
 
   RemoteApiService({
+    super.apiUrl = kApiUri,
     super.storagePath = '',
     super.userAgent = kAppTitle,
-    super.serverName = kAppLinkBase,
+    super.isDebugMode = kDebugMode,
     super.jwtExpiresIn = kJwtExpiresIn,
   });
 
@@ -26,21 +28,15 @@ class RemoteApiService extends TenturaApi {
 }
 
 extension ErrorHandler<TData, TVars> on OperationResponse<TData, TVars> {
-  OperationResponse<TData, TVars> throwIfError({
-    bool failOnNull = true,
-    String? label,
-  }) {
+  TData dataOrThrow({String? label}) {
     if (hasErrors) {
       throw GraphQLException(
         error: linkException ?? graphqlErrors,
         label: label,
       );
     }
-    if (failOnNull && data == null) {
-      throw GraphQLNoDataException(label: label);
-    }
-    return this;
-  }
+    if (data == null) throw GraphQLNoDataException(label: label);
 
-  TData dataOrThrow({String? label}) => throwIfError(label: label).data!;
+    return data!;
+  }
 }
