@@ -1,7 +1,10 @@
 sealed class BeaconException implements Exception {
   const BeaconException([this.message]);
 
-  final String? message;
+  final Object? message;
+
+  @override
+  String toString() => message?.toString() ?? super.toString();
 }
 
 final class BeaconFetchException extends BeaconException {
@@ -14,4 +17,8 @@ final class BeaconCreateException extends BeaconException {
 
 final class BeaconUpdateException extends BeaconException {
   const BeaconUpdateException([super.message]);
+}
+
+final class BeaconDeleteException extends BeaconException {
+  const BeaconDeleteException([super.message]);
 }
