@@ -3,21 +3,30 @@ import 'package:flutter/material.dart';
 import 'package:auto_route/auto_route.dart';
 
 import 'package:tentura/consts.dart';
-import 'package:tentura/domain/entity/geo.dart';
 import 'package:tentura/ui/utils/ui_utils.dart';
-import 'package:tentura/ui/dialog/choose_location_dialog.dart';
-
-import 'package:tentura/features/context/ui/widget/context_drop_down.dart';
 import 'package:tentura/ui/widget/tentura_icons.dart';
 
+import 'package:tentura/features/geo/domain/entity/coordinates.dart';
+import 'package:tentura/features/geo/ui/dialog/choose_location_dialog.dart';
+import 'package:tentura/features/context/ui/widget/context_drop_down.dart';
+import 'package:tentura/features/context/ui/bloc/context_cubit.dart';
+
+import '../../domain/entity/beacon.dart';
 import '../bloc/beacon_cubit.dart';
+import '../dialog/beacon_publish_dialog.dart';
 
 @RoutePage()
-class BeaconCreateScreen extends StatefulWidget {
+class BeaconCreateScreen extends StatefulWidget implements AutoRouteWrapper {
   const BeaconCreateScreen({super.key});
 
   @override
   State<BeaconCreateScreen> createState() => _BeaconCreateScreenState();
+
+  @override
+  Widget wrappedRoute(BuildContext context) => BlocProvider(
+        create: (_) => ContextCubit(),
+        child: this,
+      );
 }
 
 class _BeaconCreateScreenState extends State<BeaconCreateScreen> {
@@ -28,12 +37,9 @@ class _BeaconCreateScreenState extends State<BeaconCreateScreen> {
   final _dateRangeController = TextEditingController();
   final _descriptionController = TextEditingController();
 
-  late final _beaconCubit = context.read<BeaconCubit>();
-
   DateTimeRange? _dateRange;
   Coordinates? _coordinates;
   Uint8List? _image;
-  String? _context;
 
   @override
   void dispose() {
@@ -50,39 +56,13 @@ class _BeaconCreateScreenState extends State<BeaconCreateScreen> {
         appBar: AppBar(
           actions: [
             // Publish Button
-            TextButton(
-              onPressed: () async {
-                if (_formKey.currentState?.validate() ?? false) {
-                  try {
-                    await _beaconCubit.create(
-                      title: _titleController.text,
-                      description: _descriptionController.text,
-                      coordinates: _coordinates,
-                      dateRange: _dateRange,
-                      context: _context,
-                      image: _image,
-                    );
-                    if (context.mounted) await context.maybePop();
-                  } catch (e) {
-                    if (context.mounted) {
-                      showSnackBar(
-                        context,
-                        isError: true,
-                        text: e.toString(),
-                      );
-                    }
-                  }
-                } else {
-                  showSnackBar(
-                    context,
-                    isError: true,
-                    text: 'Please check input data!',
-                  );
-                }
-              },
-              child: const Text('Publish'),
+            Padding(
+              padding: kPaddingH,
+              child: TextButton(
+                onPressed: _onPublish,
+                child: const Text('Publish'),
+              ),
             ),
-            const SizedBox(width: 16),
           ],
         ),
 
@@ -102,13 +82,8 @@ class _BeaconCreateScreenState extends State<BeaconCreateScreen> {
                 ),
                 keyboardType: TextInputType.text,
                 maxLength: kTitleMaxLength,
-                onTapOutside: (event) => FocusScope.of(context).unfocus(),
-                validator: (value) {
-                  if (value == null || value.length < kTitleMinLength) {
-                    return 'Title too short';
-                  }
-                  return null;
-                },
+                onTapOutside: (_) => FocusScope.of(context).unfocus(),
+                validator: _titleValidator,
               ),
 
               // Description
@@ -120,127 +95,181 @@ class _BeaconCreateScreenState extends State<BeaconCreateScreen> {
                 keyboardType: TextInputType.multiline,
                 maxLength: kDescriptionLength,
                 maxLines: null,
-                onTapOutside: (event) => FocusScope.of(context).unfocus(),
+                onTapOutside: (_) => FocusScope.of(context).unfocus(),
               ),
 
               // Context
-              ContextDropDown(
-                onChanged: (value) async => _context = value,
+              const Padding(
+                padding: kPaddingSmallV,
+                child: ContextDropDown(key: Key('BeaconContextSelector')),
               ),
 
               // Location
-              const SizedBox(height: 20),
-              TextFormField(
-                readOnly: true,
-                controller: _locationController,
-                decoration: InputDecoration(
-                  hintText: 'Add location',
-                  suffixIcon: _locationController.text.isEmpty
-                      ? const Icon(TenturaIcons.location)
-                      : IconButton(
-                          onPressed: () {
-                            _coordinates = null;
-                            _locationController.clear();
-                          },
-                          icon: const Icon(Icons.cancel_rounded),
-                        ),
+              Padding(
+                padding: kPaddingSmallV,
+                child: TextFormField(
+                  readOnly: true,
+                  controller: _locationController,
+                  decoration: InputDecoration(
+                    hintText: 'Add location',
+                    suffixIcon: _locationController.text.isEmpty
+                        ? const Icon(TenturaIcons.location)
+                        : IconButton(
+                            icon: const Icon(Icons.cancel_rounded),
+                            onPressed: _onClearLocation,
+                          ),
+                  ),
+                  onTap: _onPickLocation,
                 ),
-                onTap: () async {
-                  final location = await ChooseLocationDialog.show(
-                    context,
-                    center: _coordinates,
-                  );
-                  if (location != null) {
-                    _coordinates = location.coords;
-                    _locationController.text = location.place.country == null
-                        ? _coordinates.toString()
-                        : '${location.place.locality ?? "Unknown"}, '
-                            '${location.place.country}';
-                  }
-                },
               ),
-              const Padding(padding: kPaddingSmallV),
 
-              // Time
-              TextFormField(
-                readOnly: true,
-                controller: _dateRangeController,
-                decoration: InputDecoration(
-                  hintText: 'Set display period',
-                  suffixIcon: _dateRangeController.text.isEmpty
-                      ? const Icon(TenturaIcons.calendar)
-                      : IconButton(
-                          onPressed: () {
-                            _dateRange = null;
-                            _dateRangeController.clear();
-                          },
-                          icon: const Icon(Icons.cancel_rounded),
-                        ),
+              // Date Range
+              Padding(
+                padding: kPaddingSmallV,
+                child: TextFormField(
+                  readOnly: true,
+                  controller: _dateRangeController,
+                  decoration: InputDecoration(
+                    hintText: 'Set display period',
+                    suffixIcon: _dateRangeController.text.isEmpty
+                        ? const Icon(TenturaIcons.calendar)
+                        : IconButton(
+                            icon: const Icon(Icons.cancel_rounded),
+                            onPressed: _onClearDate,
+                          ),
+                  ),
+                  onTap: _onPickDate,
                 ),
-                onTap: () async {
-                  final now = DateTime.timestamp();
-                  final dateRange = await showDateRangePicker(
-                    context: context,
-                    firstDate: now,
-                    lastDate: now.add(const Duration(days: 365)),
-                    initialEntryMode: DatePickerEntryMode.calendarOnly,
-                  );
-                  if (dateRange != null) {
-                    _dateRange = dateRange;
-                    _dateRangeController.text =
-                        '${fYMD(_dateRange!.start)} - ${fYMD(_dateRange!.end)}';
-                  }
-                },
               ),
-              const Padding(
-                  padding: EdgeInsets.symmetric(vertical: kSpacingSmall)),
 
-              // Image Input
-              TextFormField(
-                readOnly: true,
-                controller: _imageController,
-                decoration: InputDecoration(
-                  hintText: 'Attach image',
-                  suffixIcon: _image == null
-                      ? const Icon(Icons.add_a_photo_rounded)
-                      : IconButton(
-                          onPressed: () {
-                            _imageController.clear();
-                            setState(() => _image = null);
-                          },
-                          icon: const Icon(Icons.cancel_rounded),
-                        ),
+              // Image
+              Padding(
+                padding: kPaddingSmallV,
+                child: TextFormField(
+                  readOnly: true,
+                  controller: _imageController,
+                  decoration: InputDecoration(
+                    hintText: 'Attach image',
+                    suffixIcon: _image == null
+                        ? const Icon(Icons.add_a_photo_rounded)
+                        : IconButton(
+                            icon: const Icon(Icons.cancel_rounded),
+                            onPressed: _onImageClear,
+                          ),
+                  ),
+                  onTap: _onPickImage,
                 ),
-                onTap: () async {
-                  final newImage = await _beaconCubit.pickImage();
-                  if (newImage != null) {
-                    _imageController.text = newImage.name;
-                    _image = newImage.bytes;
-                    setState(() {});
-                  }
-                },
               ),
 
               // Image Container
-              Container(
-                decoration: BoxDecoration(
-                  border:
-                      _image == null ? Border.all(color: Colors.black12) : null,
+              Padding(
+                padding: const EdgeInsets.all(48),
+                child: GestureDetector(
+                  onTap: _onPickImage,
+                  child: _image == null
+                      ? DecoratedBox(
+                          decoration: BoxDecoration(
+                            border: Border.all(color: Colors.black12),
+                          ),
+                          child: const Icon(
+                            Icons.photo_outlined,
+                            size: 256,
+                          ),
+                        )
+                      : Image.memory(
+                          _image!,
+                          key: ObjectKey(_image),
+                          fit: BoxFit.fitWidth,
+                        ),
                 ),
-                margin: const EdgeInsets.symmetric(vertical: 48),
-                child: _image == null
-                    ? const Icon(
-                        Icons.photo_outlined,
-                        size: 200,
-                      )
-                    : Image.memory(
-                        _image!,
-                        key: ObjectKey(_image),
-                        fit: BoxFit.fitWidth,
-                      ),
               ),
             ],
           ),
         ),
       );
+
+  Future<void> _onPickLocation() async {
+    final location = await ChooseLocationDialog.show(
+      context,
+      center: _coordinates,
+    );
+    if (location == null) return;
+    _locationController.text =
+        location.place?.toString() ?? _coordinates.toString();
+    setState(() => _coordinates = location.coords);
+  }
+
+  void _onClearLocation() {
+    _locationController.clear();
+    setState(() => _coordinates = null);
+  }
+
+  Future<void> _onPickDate() async {
+    final now = DateTime.timestamp();
+    final dateRange = await showDateRangePicker(
+      context: context,
+      firstDate: now,
+      currentDate: now,
+      lastDate: now.add(const Duration(days: 365)),
+      initialEntryMode: DatePickerEntryMode.calendarOnly,
+      saveText: 'Ok',
+    );
+    if (dateRange == null) return;
+    _dateRangeController.text =
+        '${fYMD(dateRange.start)} - ${fYMD(dateRange.end)}';
+    setState(() => _dateRange = dateRange);
+  }
+
+  void _onClearDate() {
+    _dateRangeController.clear();
+    setState(() => _dateRange = null);
+  }
+
+  Future<void> _onPickImage() async {
+    final newImage = await pickImage();
+    if (newImage == null) return;
+    _imageController.text = newImage.name;
+    setState(() => _image = newImage.bytes);
+  }
+
+  void _onImageClear() {
+    _imageController.clear();
+    setState(() => _image = null);
+  }
+
+  String? _titleValidator(String? title) {
+    if (title == null || title.length < kTitleMinLength) {
+      return 'Title too short';
+    }
+    return null;
+  }
+
+  Future<void> _onPublish() async {
+    if (_formKey.currentState?.validate() ?? false) {
+      final contextName = context.read<ContextCubit>().state.selected;
+      if (await BeaconPublishDialog.show(context) ?? false) {
+        try {
+          await GetIt.I<BeaconCubit>().create(
+            beacon: emptyBeacon.copyWith(
+              title: _titleController.text,
+              description: _descriptionController.text,
+              coordinates: _coordinates,
+              dateRange: _dateRange,
+              context: contextName,
+            ),
+            image: _image,
+          );
+          if (mounted) await context.maybePop();
+        } catch (e) {
+          if (mounted) {
+            showSnackBar(
+              context,
+              isError: true,
+              text: e.toString(),
+            );
+          }
+        }
+      }
+    }
+  }
 }
