@@ -2,13 +2,14 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import 'package:tentura/app/router/root_router.dart';
-import 'package:tentura/features/context/ui/bloc/context_cubit.dart';
 import 'package:tentura/ui/utils/ui_utils.dart';
 import 'package:tentura/ui/widget/beacon_image.dart';
 import 'package:tentura/ui/widget/tentura_icons.dart';
-import 'package:tentura/ui/widget/place_name_text.dart';
-import 'package:tentura/ui/dialog/choose_location_dialog.dart';
 import 'package:tentura/ui/widget/show_more_text.dart';
+
+import 'package:tentura/features/context/ui/bloc/context_cubit.dart';
+import 'package:tentura/features/geo/ui/widget/place_name_text.dart';
+import 'package:tentura/features/geo/ui/dialog/choose_location_dialog.dart';
 
 import '../../domain/entity/beacon.dart';
 
@@ -16,15 +17,17 @@ class BeaconInfo extends StatelessWidget {
   const BeaconInfo({
     required this.beacon,
     this.isTitleLarge = false,
+    this.isShowMoreEnabled = true,
     super.key,
   });
 
   final Beacon beacon;
   final bool isTitleLarge;
+  final bool isShowMoreEnabled;
 
   @override
   Widget build(BuildContext context) {
-    final textTheme = Theme.of(context).textTheme;
+    final theme = Theme.of(context);
     return Column(
       crossAxisAlignment: CrossAxisAlignment.stretch,
       children: [
@@ -59,8 +62,8 @@ class BeaconInfo extends StatelessWidget {
                   textAlign: TextAlign.left,
                   overflow: TextOverflow.ellipsis,
                   style: isTitleLarge
-                      ? textTheme.headlineLarge
-                      : textTheme.headlineMedium,
+                      ? theme.textTheme.headlineLarge
+                      : theme.textTheme.headlineMedium,
                 ),
               ),
             ],
@@ -84,71 +87,68 @@ class BeaconInfo extends StatelessWidget {
                   maxLines: 1,
                   textAlign: TextAlign.left,
                   overflow: TextOverflow.ellipsis,
-                  style: textTheme.bodySmall,
+                  style: theme.textTheme.bodySmall,
                 ),
               ],
             ),
           ),
+
         // Beacon Description
         if (beacon.description.isNotEmpty)
           Padding(
             padding: kPaddingSmallT,
-            child: ShowMoreText(
-              beacon.description,
-              style: ShowMoreText.buildTextStyle(context),
-            ),
+            child: isShowMoreEnabled
+                ? ShowMoreText(
+                    beacon.description,
+                    style: ShowMoreText.buildTextStyle(context),
+                  )
+                : Text(
+                    beacon.description,
+                    style: ShowMoreText.buildTextStyle(context),
+                  ),
           ),
 
         // Beacon Geolocation
         if (beacon.context.isNotEmpty || beacon.coordinates != null)
           Padding(
             padding: kPaddingSmallT,
-            child: Wrap(
-              runSpacing: kSpacingSmall,
-              crossAxisAlignment: WrapCrossAlignment.center,
-              alignment: WrapAlignment.spaceBetween,
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: [
                 // Beacon Geolocation
-                if (beacon.coordinates != null)
+                if (beacon.coordinates?.isNotEmpty ?? false)
                   TextButton.icon(
-                    icon: const Icon(
-                      TenturaIcons.location,
-                      size: 18,
-                    ),
-                    style: TextButton.styleFrom(
-                      padding: EdgeInsets.zero,
-                    ),
+                    icon: const Icon(TenturaIcons.location, size: 18),
+                    style: TextButton.styleFrom(padding: EdgeInsets.zero),
                     label: kIsWeb
                         ? const Text('Show on the map')
                         : PlaceNameText(
                             coords: beacon.coordinates!,
-                            style: textTheme.bodySmall,
+                            style: theme.textTheme.bodySmall,
                           ),
                     onPressed: () => ChooseLocationDialog.show(
                       context,
                       center: beacon.coordinates,
                     ),
-                  )
-                else
-                  const Spacer(),
+                  ),
+
+                const Spacer(),
+
                 // Beacon Topic
                 if (beacon.context.isNotEmpty)
                   TextButton(
                     style: const ButtonStyle(
-                      padding:
-                          WidgetStatePropertyAll<EdgeInsets>(EdgeInsets.zero),
+                      padding: WidgetStatePropertyAll(EdgeInsets.zero),
                       visualDensity: VisualDensity.compact,
                     ),
                     child: Text(
                       '#${beacon.context} ',
-                      style: textTheme.bodySmall?.copyWith(
-                          color: Theme.of(context).colorScheme.primary),
+                      style: theme.textTheme.bodySmall?.copyWith(
+                        color: theme.colorScheme.primary,
+                      ),
                     ),
                     onPressed: () async {
-                      await GetIt.I<ContextCubit>().add(
-                        contextName: beacon.context,
-                        select: false,
-                      );
+                      await GetIt.I<ContextCubit>().add(beacon.context);
                       if (context.mounted) {
                         showSnackBar(
                           context,
