@@ -1,5 +1,5 @@
 import 'package:tentura/features/comment/domain/entity/comment.dart';
-import 'package:tentura/features/profile/domain/entity/profile.dart';
+import 'package:tentura/features/profile/data/model/user_model.dart';
 
 import '../gql/_g/comment_model.data.gql.dart';
 
@@ -10,11 +10,7 @@ extension type const CommentModel(GCommentModel i) implements GCommentModel {
         beaconId: i.beacon_id,
         createdAt: i.created_at,
         myVote: i.my_vote ?? 0,
-        score: double.tryParse(i.score?.value ?? '') ?? 0,
-        author: Profile(
-          id: i.author.id,
-          title: i.author.title,
-          hasAvatar: i.author.has_picture,
-        ),
+        score: double.tryParse(i.scores?.first.dst_score?.value ?? '') ?? 0,
+        author: (i.author as UserModel).toEntity,
       );
 }
