@@ -1,9 +1,12 @@
 import 'dart:async';
+import 'package:get_it/get_it.dart';
 import 'package:injectable/injectable.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 
-import 'package:tentura/features/auth/domain/exception.dart';
+import 'package:tentura/domain/entity/repository_event.dart';
+import 'package:tentura/ui/bloc/state_base.dart';
 
+import '../../domain/exception.dart';
+import '../../domain/entity/context_entity.dart';
 import '../../domain/use_case/context_case.dart';
 import 'context_state.dart';
 
@@ -12,79 +15,101 @@ export 'package:get_it/get_it.dart';
 
 export 'context_state.dart';
 
-@singleton
+@lazySingleton
 class ContextCubit extends Cubit<ContextState> {
+  ContextCubit({
+    bool fromCache = true,
+    ContextCase? contextCase,
+  })  : _contextCase = contextCase ?? GetIt.I<ContextCase>(),
+        super(const ContextState()) {
+    _contextChanges.resume();
+    fetch(fromCache: fromCache);
+  }
+
   @factoryMethod
-  ContextCubit.current(this._contextCase)
-      : super(const ContextState(userId: '')) {
+  ContextCubit.global(this._contextCase) : super(const ContextState()) {
     _authChanges.resume();
+    _contextChanges.resume();
   }
 
   final ContextCase _contextCase;
 
   late final _authChanges = _contextCase.currentAccountChanges.listen(
     (id) async {
-      emit(ContextState(userId: id));
-      await fetch();
+      if (id.isNotEmpty) await fetch(fromCache: false);
+    },
+    cancelOnError: false,
+  );
+
+  late final _contextChanges = _contextCase.contextChanges.listen(
+    (event) => switch (event) {
+      final RepositoryEventCreate<ContextEntity> entity => emit(ContextState(
+          contexts: state.contexts..add(entity.value.name),
+          selected: state.selected,
+        )),
+      final RepositoryEventDelete<ContextEntity> entity => emit(ContextState(
+          contexts: state.contexts..remove(entity.id),
+          selected: state.selected == entity.id ? '' : state.selected,
+        )),
+      RepositoryEventUpdate<ContextEntity>() => throw UnimplementedError(),
     },
     cancelOnError: false,
     onError: (Object? e) =>
-        emit(state.setError(e ?? const AuthExceptionUnknown())),
+        emit(state.setError(e ?? const ContextUnknownException())),
   );
 
-  @override
   @disposeMethod
-  Future<void> close() async {
+  Future<void> dispose() async {
     await _authChanges.cancel();
-    return super.close();
+    await _contextChanges.cancel();
+    return close();
   }
 
-  Future<void> fetch() async {
+  Future<void> fetch({bool fromCache = true}) async {
     emit(state.setLoading());
     try {
-      emit(state.copyWith(contexts: (await _contextCase.fetch()).toSet()));
+      final contexts = await _contextCase.fetch(fromCache: fromCache);
+      emit(ContextState(
+        contexts: contexts.toSet(),
+        selected: state.selected,
+      ));
     } catch (e) {
       emit(state.setError(e));
     }
   }
 
   String select(String contextName) {
-    emit(state.copyWith(selected: contextName));
+    emit(ContextState(
+      contexts: state.contexts,
+      selected: contextName,
+    ));
     return contextName;
   }
 
-  Future<void> add({
-    required String contextName,
-    required bool select,
+  Future<void> add(
+    String contextName, {
+    bool select = true,
   }) async {
     if (state.contexts.contains(contextName)) return;
+
+    emit(state.setLoading());
     try {
-      final context = await _contextCase.add(contextName);
-      emit(state.copyWith(
-        selected: context,
-        contexts: {context, ...state.contexts},
-      ));
+      await _contextCase.add(contextName);
+      if (select) emit(state.copyWith(selected: contextName));
     } catch (e) {
       emit(state.setError(e));
     }
   }
 
-  /// Returns `true` if current context deleted
-  Future<bool> delete(String contextName) async {
-    final isCurrent = state.selected == contextName;
+  Future<void> delete(String contextName) async {
+    emit(state.setLoading());
     try {
-      await _contextCase.delete(
-        contextName: contextName,
-        userId: state.userId,
-      );
-      emit(state.copyWith(
-        selected: isCurrent ? '' : state.selected,
-        contexts: state.contexts.where((e) => e != contextName).toSet(),
-      ));
-      return isCurrent;
+      await _contextCase.delete(contextName);
+      if (contextName == state.selected) {
+        emit(state.copyWith(selected: ''));
+      }
     } catch (e) {
       emit(state.setError(e));
-      return false;
     }
   }
 }
