@@ -2,10 +2,9 @@ import 'package:tentura/ui/bloc/state_base.dart';
 
 part 'context_state.freezed.dart';
 
-@freezed
+@Freezed(makeCollectionsUnmodifiable: false)
 class ContextState with _$ContextState, StateFetchMixin {
   const factory ContextState({
-    required String userId,
     @Default('') String selected,
     @Default({}) Set<String> contexts,
     @Default(FetchStatus.isSuccess) FetchStatus status,
@@ -14,7 +13,10 @@ class ContextState with _$ContextState, StateFetchMixin {
 
   const ContextState._();
 
-  ContextState setLoading() => copyWith(status: FetchStatus.isLoading);
+  ContextState setLoading() => copyWith(
+        status: FetchStatus.isLoading,
+        error: null,
+      );
 
   ContextState setError(Object error) => copyWith(
         status: FetchStatus.isFailure,
