@@ -1,7 +1,5 @@
 import 'package:flutter/material.dart';
 
-import '../bloc/context_cubit.dart';
-
 class ContextRemoveDialog extends StatelessWidget {
   static Future<bool?> show(
     BuildContext context, {
@@ -9,8 +7,7 @@ class ContextRemoveDialog extends StatelessWidget {
   }) =>
       showDialog<bool>(
         context: context,
-        useRootNavigator: false,
-        builder: (context) => ContextRemoveDialog(contextName: contextName),
+        builder: (_) => ContextRemoveDialog(contextName: contextName),
       );
 
   const ContextRemoveDialog({
@@ -26,11 +23,7 @@ class ContextRemoveDialog extends StatelessWidget {
         content: Text('Topic $contextName will be removed from your list!'),
         actions: [
           TextButton(
-            onPressed: () async {
-              final isCurrent =
-                  await GetIt.I<ContextCubit>().delete(contextName);
-              if (context.mounted) Navigator.of(context).pop(isCurrent);
-            },
+            onPressed: () => Navigator.of(context).pop(true),
             child: const Text('Yes'),
           ),
           TextButton(
