@@ -23,51 +23,49 @@ class FavoritesScreen extends StatelessWidget {
         listener: showSnackBarError,
         buildWhen: (p, c) => c.hasNoError,
         builder: (context, state) {
-          if (state.isLoading) {
-            // Loading state
-            return const Center(
-              child: CircularProgressIndicator.adaptive(),
-            );
-          } else if (state.beacons.isEmpty) {
-            // Empty state
-            return RefreshIndicator.adaptive(
-              onRefresh: favoritesCubit.fetch,
-              child: CustomScrollView(
-                slivers: [
-                  SliverFillRemaining(
-                    child: Container(
-                      alignment: Alignment.center,
-                      child: Text(
-                        'There is nothing here yet',
-                        style: Theme.of(context).textTheme.displaySmall,
-                        textAlign: TextAlign.center,
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            );
-          }
+          return state.isLoading
+              // Loading state
+              ? const Center(
+                  child: CircularProgressIndicator.adaptive(),
+                )
+              : RefreshIndicator.adaptive(
+                  onRefresh: favoritesCubit.fetch,
+                  child: state.beacons.isEmpty
 
-          // Beacons list
-          return RefreshIndicator.adaptive(
-            onRefresh: favoritesCubit.fetch,
-            child: ListView.separated(
-              key: const PageStorageKey('FavoritesListView'),
-              itemCount: state.beacons.length,
-              separatorBuilder: (_, __) => const Divider(),
-              itemBuilder: (context, i) {
-                final beacon = state.beacons[i];
-                return Padding(
-                  padding: kPaddingV,
-                  child: BeaconTile(
-                    beacon: beacon,
-                    key: ValueKey(beacon),
-                  ),
+                      // Empty state
+                      ? CustomScrollView(
+                          slivers: [
+                            SliverFillRemaining(
+                              child: Container(
+                                alignment: Alignment.center,
+                                child: Text(
+                                  'There is nothing here yet',
+                                  style:
+                                      Theme.of(context).textTheme.displaySmall,
+                                  textAlign: TextAlign.center,
+                                ),
+                              ),
+                            ),
+                          ],
+                        )
+
+                      // Beacons list
+                      : ListView.separated(
+                          key: const PageStorageKey('FavoritesListView'),
+                          itemCount: state.beacons.length,
+                          separatorBuilder: (_, __) => const Divider(),
+                          itemBuilder: (context, i) {
+                            final beacon = state.beacons[i];
+                            return Padding(
+                              padding: kPaddingV,
+                              child: BeaconTile(
+                                beacon: beacon,
+                                key: ValueKey(beacon),
+                              ),
+                            );
+                          },
+                        ),
                 );
-              },
-            ),
-          );
         },
       ),
     );
