@@ -1,9 +1,9 @@
 import 'package:injectable/injectable.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../domain/entity/likable_entity.dart';
+import 'package:tentura/domain/entity/likable.dart';
+
 import '../../domain/use_case/like_case.dart';
-import '../../domain/typedef.dart';
 import 'like_state.dart';
 
 export 'package:flutter_bloc/flutter_bloc.dart';
@@ -12,28 +12,34 @@ export 'like_state.dart';
 
 @lazySingleton
 class LikeCubit extends Cubit<LikeState> {
-  LikeCubit(this._likeCase) : super(const LikeState()) {
+  LikeCubit(this._likeCase)
+      : super(LikeState(
+          likes: {},
+          updatedAt: DateTime.timestamp(),
+        )) {
     _authChanges.resume();
     _likeChanges.resume();
   }
 
   final LikeCase _likeCase;
 
+  late final _authChanges = _likeCase.currentAccountChanges.listen(
+    (id) => emit(LikeState(
+      likes: {},
+      updatedAt: DateTime.timestamp(),
+    )),
+    cancelOnError: false,
+  );
+
   late final _likeChanges = _likeCase.likeChanges.listen(
     (e) {
-      state.likes[e.id] = e.amount;
-      emit(LikeState(likes: state.likes));
+      state.likes[e.id] = e.value.votes;
+      emit(state.copyWith(
+        updatedAt: DateTime.timestamp(),
+      ));
     },
     cancelOnError: false,
   );
-
-  late final _authChanges = _likeCase.currentAccountChanges.listen(
-    // ignore: prefer_const_constructors
-    (id) => emit(LikeState(likes: {})),
-    cancelOnError: false,
-  );
-
-  Stream<LikeAmount> get likeChanges => _likeCase.likeChanges;
 
   @override
   @disposeMethod
@@ -43,13 +49,21 @@ class LikeCubit extends Cubit<LikeState> {
     return super.close();
   }
 
-  Future<void> addLikeAmount({
-    required LikableEntity entity,
-    required int amount,
-  }) async {
+  Future<void> incrementLike(Likable entity) async {
     try {
-      await _likeCase.addLikeAmount(
-        amount: state.getLikeAmount(entity).amount + amount,
+      await _likeCase.setLikeAmount(
+        amount: (state.likes[entity.id] ?? entity.votes) + 1,
+        entity: entity,
+      );
+    } catch (e) {
+      emit(state.setError(e));
+    }
+  }
+
+  Future<void> decrementLike(Likable entity) async {
+    try {
+      await _likeCase.setLikeAmount(
+        amount: (state.likes[entity.id] ?? entity.votes) - 1,
         entity: entity,
       );
     } catch (e) {
