@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
 
+import 'package:tentura/domain/entity/likable.dart';
 import 'package:tentura/ui/widget/tentura_icons.dart';
 
-import '../../domain/entity/likable_entity.dart';
-import '../../domain/typedef.dart';
 import '../bloc/like_cubit.dart';
 
 class LikeControl extends StatelessWidget {
@@ -13,7 +12,7 @@ class LikeControl extends StatelessWidget {
     super.key,
   });
 
-  final LikableEntity entity;
+  final Likable entity;
 
   @override
   Widget build(BuildContext context) {
@@ -29,22 +28,20 @@ class LikeControl extends StatelessWidget {
         children: [
           IconButton(
             icon: const Icon(TenturaIcons.arrowUp),
-            onPressed: () async =>
-                likeCubit.addLikeAmount(entity: entity, amount: 1),
+            onPressed: () => likeCubit.incrementLike(entity),
           ),
-          StreamBuilder<LikeAmount>(
+          BlocSelector<LikeCubit, LikeState, int>(
             key: ValueKey(entity),
-            initialData: likeCubit.state.getLikeAmount(entity),
-            stream: likeCubit.likeChanges.where((e) => e.id == entity.id),
-            builder: (context, snapshot) => Text(
-              (snapshot.data?.amount ?? 0).toString(),
+            bloc: likeCubit,
+            selector: (state) => state.likes[entity.id] ?? entity.votes,
+            builder: (context, state) => Text(
+              state.toString(),
               style: theme.textTheme.bodyMedium,
             ),
           ),
           IconButton(
             icon: const Icon(TenturaIcons.arrowDown),
-            onPressed: () async =>
-                likeCubit.addLikeAmount(entity: entity, amount: -1),
+            onPressed: () => likeCubit.decrementLike(entity),
           ),
         ],
       ),
