@@ -16,7 +16,11 @@ class MyFieldRepository {
   final RemoteApiService _remoteApiService;
 
   Future<Iterable<Beacon>> fetch({required String context}) => _remoteApiService
-      .request(GMyFieldFetchReq((r) => r.vars.context = context))
+      .request(GMyFieldFetchReq((r) => r
+        ..context = const Context().withEntry(HttpLinkHeaders(headers: {
+          headerQueryContext: context,
+        }))
+        ..vars.context = context))
       .firstWhere((e) => e.dataSource == DataSource.Link)
       .then((r) => r.dataOrThrow(label: _label).my_field.nonNulls)
       .then((v) => v.map((e) => (e.beacon! as BeaconModel).toEntity));
