@@ -5,6 +5,7 @@ import 'package:tentura/ui/utils/ui_utils.dart';
 
 import 'package:tentura/features/auth/ui/bloc/auth_cubit.dart';
 import 'package:tentura/features/beacon/ui/widget/beacon_tile.dart';
+import 'package:tentura/features/context/ui/bloc/context_cubit.dart';
 import 'package:tentura/features/context/ui/widget/context_drop_down.dart';
 
 import '../bloc/my_field_cubit.dart';
@@ -18,25 +19,34 @@ class MyFieldScreen extends StatelessWidget implements AutoRouteWrapper {
       BlocSelector<AuthCubit, AuthState, String>(
         bloc: GetIt.I<AuthCubit>(),
         selector: (state) => state.currentAccountId,
-        builder: (context, accountId) => BlocProvider(
-          create: (_) => MyFieldCubit(),
-          child: this,
+        builder: (context, accountId) => MultiBlocProvider(
+          key: ValueKey(accountId),
+          providers: [
+            BlocProvider.value(
+              value: GetIt.I<ContextCubit>(),
+            ),
+            BlocProvider(
+              create: (context) => MyFieldCubit(
+                initialContext: GetIt.I<ContextCubit>().state.selected,
+              ),
+            ),
+          ],
+          child: BlocListener<ContextCubit, ContextState>(
+            listenWhen: (p, c) => p.selected != c.selected,
+            listener: (context, state) =>
+                context.read<MyFieldCubit>().fetch(state.selected),
+            child: this,
+          ),
         ),
       );
 
   @override
   Widget build(BuildContext context) => SafeArea(
-        minimum: const EdgeInsets.only(
-          left: kSpacingMedium,
-          right: kSpacingMedium,
-          top: kSpacingMedium,
-        ),
+        minimum: kPaddingAll,
         child: Column(
           children: [
             // Context selector
-            ContextDropDown(
-              onChanged: context.read<MyFieldCubit>().fetch,
-            ),
+            const ContextDropDown(key: Key('MyFieldContextSelector')),
 
             // Beacons list
             Expanded(
@@ -45,6 +55,11 @@ class MyFieldScreen extends StatelessWidget implements AutoRouteWrapper {
                 listener: showSnackBarError,
                 buildWhen: (p, c) => c.hasNoError,
                 builder: (context, state) {
+                  if (state.isLoading) {
+                    return const Center(
+                      child: CircularProgressIndicator.adaptive(),
+                    );
+                  }
                   return ListView.separated(
                     itemCount: state.beacons.length,
                     separatorBuilder: (_, __) => const Divider(),
