@@ -9,14 +9,14 @@ import '../model/profile_model.dart';
 @singleton
 class ProfileLocalRepository {
   static const _repositoryKey = 'Profile:';
-  static const _progileKey = '${_repositoryKey}Id:';
+  static const _profileKey = '${_repositoryKey}Id:';
 
   ProfileLocalRepository(this._localStorage);
 
   final LocalSecureStorage _localStorage;
 
   Future<Profile?> getProfileById(String id) => _localStorage
-      .read('$_progileKey$id')
+      .read('$_profileKey$id')
       .then((v) => switch (jsonDecode(v ?? 'null')) {
             final Map<String, dynamic> j => ProfileModel.fromJson(j).toEntity,
             _ => null,
@@ -24,11 +24,11 @@ class ProfileLocalRepository {
 
   Future<Profile> setProfile(Profile profile) => _localStorage
       .write(
-        '$_progileKey${profile.id}',
+        '$_profileKey${profile.id}',
         jsonEncode(ProfileModel.fromEntity(profile).toJson()),
       )
       .then((_) => profile);
 
   Future<String> deleteProfileById(String id) =>
-      _localStorage.write('$_progileKey$id', null).then((_) => id);
+      _localStorage.write('$_profileKey$id', null).then((_) => id);
 }
