@@ -8,7 +8,7 @@ import '../gql/_g/user_delete_by_id.req.gql.dart';
 import '../gql/_g/user_fetch_by_id.req.gql.dart';
 import '../gql/_g/user_update.req.gql.dart';
 
-@singleton
+@lazySingleton
 class ProfileRemoteRepository {
   static const _label = 'Profile';
 
