@@ -6,6 +6,7 @@ import 'package:tentura/ui/widget/avatar_image.dart';
 import 'package:tentura/ui/widget/tentura_icons.dart';
 import 'package:tentura/ui/widget/show_more_text.dart';
 import 'package:tentura/ui/widget/gradient_stack.dart';
+import 'package:tentura/ui/widget/deep_back_button.dart';
 import 'package:tentura/ui/widget/avatar_positioned.dart';
 import 'package:tentura/ui/widget/share_code_icon_button.dart';
 
@@ -44,133 +45,134 @@ class ProfileViewScreen extends StatelessWidget implements AutoRouteWrapper {
         }
         final profile = state.profile;
         final beacons = state.beacons;
-        final theme = Theme.of(context);
-        return CustomScrollView(
-          slivers: [
-            // Header
-            SliverAppBar(
-              actions: [
-                // Graph View
-                IconButton(
-                  icon: const Icon(TenturaIcons.graph),
-                  onPressed: () =>
-                      context.pushRoute(GraphRoute(focus: profile.id)),
-                ),
+        final textTheme = Theme.of(context).textTheme;
+        return Scaffold(
+          body: CustomScrollView(
+            slivers: [
+              // Header
+              SliverAppBar(
+                actions: [
+                  // Graph View
+                  IconButton(
+                    icon: const Icon(TenturaIcons.graph),
+                    onPressed: () =>
+                        context.pushRoute(GraphRoute(focus: profile.id)),
+                  ),
 
-                // Share
-                ShareCodeIconButton.id(profile.id),
+                  // Share
+                  ShareCodeIconButton.id(profile.id),
 
-                // More
-                PopupMenuButton(
-                  itemBuilder: (context) => <PopupMenuEntry<void>>[
-                    if (profile.isFriend)
-                      PopupMenuItem(
-                        onTap: profileViewCubit.removeFriend,
-                        child: const Text('Remove from my field'),
-                      )
-                    else
-                      PopupMenuItem(
-                        onTap: profileViewCubit.addFriend,
-                        child: const Text('Add to my field'),
+                  // More
+                  PopupMenuButton(
+                    itemBuilder: (context) => <PopupMenuEntry<void>>[
+                      if (profile.isFriend)
+                        PopupMenuItem(
+                          onTap: profileViewCubit.removeFriend,
+                          child: const Text('Remove from my field'),
+                        )
+                      else
+                        PopupMenuItem(
+                          onTap: profileViewCubit.addFriend,
+                          child: const Text('Add to my field'),
+                        ),
+                    ],
+                  ),
+                ],
+                floating: true,
+                leading: const DeepBackButton(),
+                expandedHeight: GradientStack.defaultHeight,
+
+                // Avatar
+                flexibleSpace: FlexibleSpaceBar(
+                  background: GradientStack(
+                    children: [
+                      AvatarPositioned(
+                        child: AvatarImage(
+                          userId: profile.imageId,
+                          size: AvatarPositioned.childSize,
+                        ),
                       ),
-                  ],
-                  useRootNavigator: true,
-                ),
-              ],
-              floating: true,
-              expandedHeight: GradientStack.defaultHeight,
-              leading: const AutoLeadingButton(),
-
-              // Avatar
-              flexibleSpace: FlexibleSpaceBar(
-                background: GradientStack(
-                  children: [
-                    AvatarPositioned(
-                      child: AvatarImage(
-                        userId: profile.imageId,
-                        size: AvatarPositioned.childSize,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-
-            // Body
-            SliverToBoxAdapter(
-              child: Padding(
-                padding: kPaddingH,
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    // Title
-                    Text(
-                      profile.title.isEmpty ? 'No name' : profile.title,
-                      textAlign: TextAlign.left,
-                      style: theme.textTheme.headlineLarge,
-                    ),
-                    const Padding(padding: kPaddingSmallV),
-
-                    // Description
-                    ShowMoreText(
-                      profile.description,
-                      style: ShowMoreText.buildTextStyle(context),
-                    ),
-                    const Divider(),
-
-                    const Padding(padding: kPaddingSmallT),
-
-                    Text(
-                      'Beacons',
-                      textAlign: TextAlign.left,
-                      style: theme.textTheme.titleLarge,
-                    ),
-                  ],
-                ),
-              ),
-            ),
-
-            // Beacons
-            if (beacons.isEmpty)
-              SliverToBoxAdapter(
-                child: Padding(
-                  padding: kPaddingAll,
-                  child: Text(
-                    'There are no beacons yet',
-                    style: theme.textTheme.bodyMedium,
+                    ],
                   ),
                 ),
-              )
-            else
-              SliverList.separated(
-                key: ValueKey(beacons),
-                itemCount: beacons.length,
-                itemBuilder: (context, i) {
-                  final beacon = beacons[i];
-                  return Padding(
+              ),
+
+              // Body
+              SliverToBoxAdapter(
+                child: Padding(
+                  padding: kPaddingH,
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      // Title
+                      Text(
+                        profile.title.isEmpty ? 'No name' : profile.title,
+                        textAlign: TextAlign.left,
+                        style: textTheme.headlineLarge,
+                      ),
+                      const Padding(padding: kPaddingSmallV),
+
+                      // Description
+                      ShowMoreText(
+                        profile.description,
+                        style: ShowMoreText.buildTextStyle(context),
+                      ),
+                      const Divider(),
+
+                      const Padding(padding: kPaddingSmallT),
+
+                      Text(
+                        'Beacons',
+                        textAlign: TextAlign.left,
+                        style: textTheme.titleLarge,
+                      ),
+                    ],
+                  ),
+                ),
+              ),
+
+              // Beacons
+              if (beacons.isEmpty)
+                SliverToBoxAdapter(
+                  child: Padding(
                     padding: kPaddingAll,
-                    child: BeaconTile(
-                      beacon: beacon,
-                      key: ValueKey(beacon),
+                    child: Text(
+                      'There are no beacons yet',
+                      style: textTheme.bodyMedium,
                     ),
-                  );
-                },
-                separatorBuilder: (_, __) =>
-                    const Divider(endIndent: 20, indent: 20),
-              ),
+                  ),
+                )
+              else
+                SliverList.separated(
+                  key: ValueKey(beacons),
+                  itemCount: beacons.length,
+                  itemBuilder: (context, i) {
+                    final beacon = beacons[i];
+                    return Padding(
+                      padding: kPaddingAll,
+                      child: BeaconTile(
+                        beacon: beacon,
+                        key: ValueKey(beacon),
+                      ),
+                    );
+                  },
+                  separatorBuilder: (_, __) =>
+                      const Divider(endIndent: 20, indent: 20),
+                ),
 
-            // Show more
-            if (beacons.isNotEmpty && state.hasNotReachedMax)
-              SliverToBoxAdapter(
-                child: Padding(
-                  padding: kPaddingAll,
-                  child: TextButton(
-                    onPressed: profileViewCubit.fetchMore,
-                    child: const Text('Show more'),
+              // Show more
+              if (beacons.isNotEmpty && state.hasNotReachedMax)
+                SliverToBoxAdapter(
+                  child: Padding(
+                    padding: kPaddingAll,
+                    child: TextButton(
+                      onPressed: profileViewCubit.fetchMore,
+                      child: const Text('Show more'),
+                    ),
                   ),
                 ),
-              ),
-          ],
+            ],
+          ),
         );
       },
     );
