@@ -4,10 +4,9 @@ import 'package:tentura/data/service/remote_api_service.dart';
 
 import 'package:tentura/features/profile/domain/entity/profile.dart';
 
-import '../domain/entity/user_rating.dart';
 import 'gql/_g/rating_fetch.req.gql.dart';
 
-@singleton
+@lazySingleton
 class RatingRepository {
   static const _label = 'Rating';
 
@@ -15,22 +14,25 @@ class RatingRepository {
 
   final RemoteApiService _remoteApiService;
 
-  Future<Iterable<UserRating>> fetch({required String context}) =>
+  Future<Iterable<Profile>> fetch({required String context}) =>
       _remoteApiService
-          .request(GRatingFetchReq((r) => r.vars.context = context))
+          .request(GRatingFetchReq((r) => r
+            ..context = const Context().withEntry(HttpLinkHeaders(headers: {
+              headerQueryContext: context,
+            }))
+            ..vars.context = context))
           .firstWhere((e) => e.dataSource == DataSource.Link)
           .then((r) => r.dataOrThrow(label: _label).rating)
           .then(
-            (r) => r.map((e) => UserRating(
-                  egoScore: double.parse(e.src_score!.value),
-                  userScore: double.parse(e.dst_score!.value),
-                  profile: Profile(
-                    id: e.user?.id ?? '',
-                    title: e.user?.title ?? '',
-                    myVote: e.user?.my_vote ?? 0,
-                    hasAvatar: e.user?.has_picture ?? false,
-                    score: double.tryParse(e.user?.score?.value ?? '') ?? 0,
-                  ),
-                )),
+            (r) => r.map(
+              (e) => Profile(
+                id: e.user?.id ?? '',
+                title: e.user?.title ?? '',
+                myVote: e.user?.my_vote ?? 0,
+                hasAvatar: e.user?.has_picture ?? false,
+                score: double.tryParse(e.dst_score?.value ?? '') ?? 0,
+                rScore: double.tryParse(e.src_score?.value ?? '') ?? 0,
+              ),
+            ),
           );
 }
