@@ -4,7 +4,6 @@ import 'package:get_it/get_it.dart';
 import 'package:tentura/ui/bloc/state_base.dart';
 
 import '../../data/rating_repository.dart';
-import '../../domain/entity/user_rating.dart';
 import 'rating_state.dart';
 
 export 'package:flutter_bloc/flutter_bloc.dart';
@@ -13,26 +12,24 @@ export 'rating_state.dart';
 
 class RatingCubit extends Cubit<RatingState> {
   RatingCubit({
+    String initialContext = '',
     RatingRepository? repository,
   })  : _repository = repository ?? GetIt.I<RatingRepository>(),
         super(const RatingState()) {
-    fetch();
+    fetch(initialContext);
   }
 
   final RatingRepository _repository;
 
-  List<UserRating> _items = [];
-
-  Future<void> fetch([String? contextName]) async {
+  Future<void> fetch([String contextName = '']) async {
     emit(state.setLoading());
     try {
-      final rating = await _repository.fetch(
-        context: contextName ?? state.context,
-      );
-      _items = rating.toList(growable: false);
       emit(state.copyWith(
+        error: null,
+        context: contextName,
         status: FetchStatus.isSuccess,
-        items: _items,
+        items: (await _repository.fetch(context: contextName))
+            .toList(growable: false),
       ));
       _sort();
     } catch (e) {
@@ -41,43 +38,38 @@ class RatingCubit extends Cubit<RatingState> {
   }
 
   void toggleSortingByAsc() {
-    emit(state.copyWith(isSortedByAsc: !state.isSortedByAsc));
+    emit(state.copyWith(
+      isSortedByAsc: !state.isSortedByAsc,
+    ));
     _sort();
   }
 
   void toggleSortingByEgo() {
-    emit(state.copyWith(isSortedByEgo: !state.isSortedByEgo));
+    emit(state.copyWith(
+      isSortedByEgo: !state.isSortedByEgo,
+    ));
     _sort();
   }
 
   void setSearchFilter(String filter) => emit(state.copyWith(
         searchFilter: filter,
-        items: _items
-            .where((e) =>
-                e.profile.title.toLowerCase().contains(filter.toLowerCase()))
-            .toList(),
       ));
 
   void clearSearchFilter() => emit(state.copyWith(
         searchFilter: '',
-        items: _items,
       ));
 
   void _sort() {
     if (state.isSortedByEgo) {
-      state.items.sort((a, b) {
-        final c = state.isSortedByAsc
-            ? a.egoScore - b.egoScore
-            : b.egoScore - a.egoScore;
-        return (c * 10000).toInt();
-      });
+      state.items.sort((a, b) =>
+          10000 *
+          (state.isSortedByAsc ? a.rScore - b.rScore : b.rScore - a.rScore)
+              .toInt());
     } else {
-      state.items.sort((a, b) {
-        final c = state.isSortedByAsc
-            ? a.userScore - b.userScore
-            : b.userScore - a.userScore;
-        return (c * 10000).toInt();
-      });
+      state.items.sort((a, b) =>
+          10000 *
+          (state.isSortedByAsc ? a.score - b.score : b.score - a.score)
+              .toInt());
     }
   }
 }
