@@ -1,27 +1,19 @@
 import 'package:flutter/material.dart';
-import 'package:cached_network_image/cached_network_image.dart';
 
 import 'package:tentura/consts.dart';
 
-class BeaconImage extends StatelessWidget {
-  static String getBeaconUrl({
-    required String userId,
-    required String beaconId,
-    String serverName = kAppLinkBase,
-  }) =>
-      'https://$serverName/images/$userId/$beaconId.jpg';
+import 'cached_image/cached_image.dart';
 
+class BeaconImage extends StatelessWidget {
   const BeaconImage({
     required this.authorId,
     this.beaconId = '',
     this.boxFit = BoxFit.cover,
-    this.serverName = kAppLinkBase,
     this.height,
     this.width,
     super.key,
   });
 
-  final String serverName;
   final String beaconId;
   final String authorId;
   final double? height;
@@ -38,17 +30,12 @@ class BeaconImage extends StatelessWidget {
     );
     return beaconId.isEmpty || authorId.isEmpty
         ? placeholder
-        : CachedNetworkImage(
+        : CachedImage(
+            imageUrl: '$kApiUri/images/$authorId/$beaconId.jpg',
+            placeholder: placeholder,
+            boxFit: boxFit,
             height: height,
             width: width,
-            filterQuality: FilterQuality.high,
-            placeholder: (context, url) => placeholder,
-            errorWidget: (context, url, error) => placeholder,
-            imageUrl: getBeaconUrl(
-              serverName: serverName,
-              userId: authorId,
-              beaconId: beaconId,
-            ),
           );
   }
 }
