@@ -8,7 +8,7 @@ import 'token_service_base.dart';
 /// Set KeyPair before fetchJWT
 class TokenService extends TokenServiceBase {
   TokenService({
-    required super.serverName,
+    required super.apiUrl,
     super.jwtExpiresIn = const Duration(minutes: 1),
   });
 
